+++ src/esp_logger/logger.cpp
@@ -145,21 +145,28 @@ bool ESPLogger::init(const LoggerConfig &config) {
 		_logLevel = _config.consoleLogLevel;
 	}
 
-	_running = false;
+		_running = false;
 	const bool shouldCreateTask = _config.enableSyncTask && _config.syncIntervalMS > 0;
+	ESPWorker::Config workerConfig{};
+	workerConfig.maxWorkers = 1;
+	workerConfig.stackSizeBytes = _config.stackSize;
+	workerConfig.priority = _config.priority;
+	workerConfig.coreId = _config.coreId;
+	workerConfig.enableExternalStacks = true;
+	_worker.init(workerConfig);
 
 	if (shouldCreateTask) {
 		_running = true;
-		BaseType_t result = xTaskCreatePinnedToCore(
-			syncTaskThunk,
-			kSyncTaskName,
-			_config.stackSize,
-			this,
-			_config.priority,
-			&_syncTask,
-			_config.coreId);
+		WorkerConfig taskConfig{};
+		taskConfig.name = kSyncTaskName;
+		taskConfig.stackSizeBytes = _config.stackSize;
+		taskConfig.priority = _config.priority;
+		taskConfig.coreId = _config.coreId;
+		WorkerResult result =
+			_config.usePSRAMBuffers ? _worker.spawnExt([this]() { syncTaskLoop(); }, taskConfig)
+			                        : _worker.spawn([this]() { syncTaskLoop(); }, taskConfig);
 
-		if (result != pdPASS) {
+		if (!result) {
 			_running = false;
 			LockGuard guard(_mutex);
 			_logs = InternalLogDeque(_logAllocator);
@@ -168,13 +175,14 @@ bool ESPLogger::init(const LoggerConfig &config) {
 			_usePSRAMBuffers = false;
 			_logAllocator = LoggerAllocator<Log>(_usePSRAMBuffers);
 			_charAllocator = LoggerAllocator<char>(_usePSRAMBuffers);
-			vSemaphoreDelete(_mutex);
-			_mutex = nullptr;
-			_syncTask = nullptr;
-			return false;
-		}
+				vSemaphoreDelete(_mutex);
+				_mutex = nullptr;
+				_syncTask.reset();
+				return false;
+			}
+		_syncTask = result.handler;
 
-	}
+		}
 
 	_initialized = true;
 	return true;
@@ -188,8 +196,9 @@ void ESPLogger::deinit() {
 	_running = false;
 
 	if (_syncTask != nullptr) {
-		vTaskDelete(_syncTask);
-		_syncTask = nullptr;
+		(void)_syncTask->wait(pdMS_TO_TICKS(200));
+		(void)_syncTask->destroy();
+		_syncTask.reset();
 	}
 
 	performSync();
@@ -210,6 +219,7 @@ void ESPLogger::deinit() {
 		vSemaphoreDelete(_mutex);
 		_mutex = nullptr;
 	}
+	_worker.deinit();
 
 	_initialized = false;
 }
@@ -441,7 +451,6 @@ void ESPLogger::syncTaskThunk(void *arg) {
 	if (instance != nullptr) {
 		instance->syncTaskLoop();
 	}
-	vTaskDelete(nullptr);
 }
 
 void ESPLogger::syncTaskLoop() {