+++ src/esp_logger/logger.h
@@ -5,9 +5,11 @@
 #include <ctime>
 #include <deque>
 #include <functional>
+#include <memory>
 #include <string>
 #include <vector>
 
+#include <ESPWorker.h>
 #include <freertos/FreeRTOS.h>
 #include <freertos/semphr.h>
 #include <freertos/task.h>
@@ -68,9 +70,10 @@ class ESPLogger {
     void syncTaskLoop();
 
     LoggerConfig _config{};
+    ESPWorker _worker{};
     bool _initialized = false;
     bool _running = false;
-    TaskHandle_t _syncTask = nullptr;
+    std::shared_ptr<WorkerHandler> _syncTask{};
     SemaphoreHandle_t _mutex = nullptr;
     InternalLogDeque _logs;
     SyncCallback _syncCallback;